Packaging (mostly) overhaul
Now packaging is done using normal setuptools and Nix. I am not sure
if pip can handle this, but it should. I don't use pip anyway.

This is now a valid Nix package that will also install a D-Bus policy
file, allowing it to claim a name on the bus.